+++ Controllers/TimesController.cs
@@ -1,12 +1,10 @@
-﻿using System;
+﻿using LigaTabajara.Models;
+using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
-using LigaTabajara.Models;
 
 namespace LigaTabajara.Controllers
 {
@@ -17,7 +15,8 @@ namespace LigaTabajara.Controllers
         // GET: Times
         public ActionResult Index()
         {
-            return View(db.Times.ToList());
+            var times = db.Times.Include(t => t.Jogadores).Include(t => t.ComissaoTecnicas).ToList();
+            return View(times);
         }
 
         // GET: Times/Details/5
@@ -27,7 +26,10 @@ namespace LigaTabajara.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Time time = db.Times.Find(id);
+            Time time = db.Times
+                .Include(t => t.Jogadores)           // Include Jogadores
+                .Include(t => t.ComissaoTecnicas)    // Include ComissaoTecnicas
+                .FirstOrDefault(t => t.Id == id);    // Use FirstOrDefault instead of Find
             if (time == null)
             {
                 return HttpNotFound();
@@ -38,46 +40,34 @@ namespace LigaTabajara.Controllers
         // GET: Times/Create
         public ActionResult Create()
         {
-            // Definir o Status como "Inativo" por padrão
-            var time = new Time
-            {
-                Status = "Inativo" // Aqui definimos o valor padrão como "Inativo"
-            };
-
-            ViewBag.CoresUniforme = Enum.GetValues(typeof(CorUniforme))
-                .Cast<CorUniforme>()
-                .Select(c => new SelectListItem
+            // Populate ViewBag.CorUniformeList with the enum values
+            ViewBag.CorUniformeList = new SelectList(
+                Enum.GetValues(typeof(CorUniforme)).Cast<CorUniforme>().Select(c => new SelectListItem
                 {
                     Value = ((int)c).ToString(),
                     Text = c.ToString()
-                });
+                }),
+                "Value",
+                "Text"
+            );
 
-            return View(time);
+            return View();
         }
 
         // POST: Times/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create(Time time)
+        public ActionResult Create([Bind(Include = "Id,Nome,Estado,AnoFundacao,Estadio,CapacidadeEstadio,CorUniforme,Status")] Time time)
         {
             if (ModelState.IsValid)
             {
+                time.Jogadores = new List<Jogador>(); // Inicializa as coleções
+                time.ComissaoTecnicas = new List<ComissaoTecnica>();
+                time.AtualizarStatus(); // Atualiza o status antes de salvar
                 db.Times.Add(time);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
-            // Repassa a lista em caso de erro
-            ViewBag.CoresUniforme = Enum.GetValues(typeof(CorUniforme))
-                .Cast<CorUniforme>()
-                .Select(c => new SelectListItem
-                {
-                    Value = ((int)c).ToString(),
-                    Text = c.ToString()
-                });
-
             return View(time);
         }
 
@@ -93,19 +83,55 @@ namespace LigaTabajara.Controllers
             {
                 return HttpNotFound();
             }
+
+            // Handle invalid CorUniforme values (e.g., 0)
+            if ((int)time.CorUniforme == 0)
+            {
+                time.CorUniforme = CorUniforme.PRIMARIA; // Default to PRIMARIA
+            }
+
+            // Populate ViewBag.CorUniformeList with the enum values
+            ViewBag.CorUniformeList = new SelectList(
+                Enum.GetValues(typeof(CorUniforme)).Cast<CorUniforme>().Select(c => new SelectListItem
+                {
+                    Value = ((int)c).ToString(),
+                    Text = c.ToString()
+                }),
+                "Value",
+                "Text",
+                (int)time.CorUniforme
+            );
+
             return View(time);
         }
 
         // POST: Times/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Nome,Estado,AnoFundacao,Estadio,CapacidadeEstadio,CorPrimariaString,CorSecundariaString,Status")] Time time)
+        public ActionResult Edit([Bind(Include = "Id,Nome,Estado,AnoFundacao,Estadio,CapacidadeEstadio,CorUniforme,Status")] Time time)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(time).State = EntityState.Modified;
+                var timeToUpdate = db.Times
+                    .Include(t => t.Jogadores)
+                    .Include(t => t.ComissaoTecnicas)
+                    .SingleOrDefault(t => t.Id == time.Id);
+
+                if (timeToUpdate == null)
+                {
+                    return HttpNotFound();
+                }
+
+                // Atualizar as propriedades manualmente
+                timeToUpdate.Nome = time.Nome;
+                timeToUpdate.Estado = time.Estado;
+                timeToUpdate.AnoFundacao = time.AnoFundacao;
+                timeToUpdate.Estadio = time.Estadio;
+                timeToUpdate.CapacidadeEstadio = time.CapacidadeEstadio;
+                timeToUpdate.CorUniforme = time.CorUniforme;
+                timeToUpdate.AtualizarStatus(); // Atualiza o status antes de salvar
+
+                db.Entry(timeToUpdate).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
@@ -132,7 +158,16 @@ namespace LigaTabajara.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Time time = db.Times.Find(id);
+            Time time = db.Times
+                .Include(t => t.Jogadores)
+                .Include(t => t.ComissaoTecnicas)
+                .SingleOrDefault(t => t.Id == id);
+
+            if (time == null)
+            {
+                return HttpNotFound();
+            }
+
             db.Times.Remove(time);
             db.SaveChanges();
             return RedirectToAction("Index");
